backend/requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.13
lxml==6.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.12.1
python-dotenv==1.0.1
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
